requirements.txt
flask
pywebview
qrcode
pillow
